fix(lysis): widen program-global heap-slot index space to u32
The heap-slot bump allocator (Walker.heap_alloc) and every field it
feeds — StoreHeap.slot, LoadHeap.slot, EmitWitnessCallHeap.out_slots,
InputSrc::Slot, the v2 header's heap_size_hint, and ssa_to_heap —
were u16. The number of distinct spilled cold vars scales with
circuit size, so a >1.5M-constraint circuit spills past 65535; the
allocator's saturating_add then clamps, aliasing every further
distinct var onto slot 65535 (a silent slot-aliasing miscompile) and
under-sizing the executor heap, which trips the heap-slot-bounds
rule at instantiate.

Widen the whole heap-slot index space u16->u32 end-to-end: allocator,
opcode fields, wire encode/decode (2->4 LE bytes), header layout
(v2 18->20 bytes), builder setters/emit helpers, and the executor's
heap sizing. The u16 input/output count prefixes on
EmitWitnessCallHeap stay: call arity is bounded by the Artik program,
not by circuit size. No bytecode VERSION bump — lysis bytecode is
never persisted or cross-version-decoded (every encode is followed by
a decode in the same process call); the rationale is documented on
the header VERSION constant. Executed semantics are unchanged
(constraints byte-identical); only the ephemeral bytecode byte width
differs where a slot is encoded.

Pinned, layered and confound-free:
- heap_slots_stay_distinct_when_spill_count_exceeds_u16_ceiling
  (walker axis): 70000 distinct cold-var spills get 70000 distinct
  slots and a matching heap_size_hint, not the u16-saturated 65535.
- heap_slot_above_u16_ceiling_round_trips_the_right_value
  (wire+executor axis): a slot past the old ceiling stores and loads
  its true value through encode/decode/execute and is not wrapped.
The v2 header test that asserted the old u16 maximum is rewritten to
assert values above the old ceiling round-trip.

+++ ir-forge/tests/walker_spill_proptest.rs
@@ -67,7 +67,7 @@ proptest! {
         let walker = Walker::<Bn254Fr>::new(FieldFamily::BnLike256);
         let Ok(program) = walker.lower(body.clone()) else { return Ok(()); };
 
-        let mut slot_writes: std::collections::HashMap<u16, u32> =
+        let mut slot_writes: std::collections::HashMap<u32, u32> =
             std::collections::HashMap::new();
         for instr in &program.body {
             if let Opcode::StoreHeap { slot, .. } = instr.opcode {
@@ -98,7 +98,7 @@ proptest! {
 
         let store_count = count_opcodes(&program, |op| matches!(op, Opcode::StoreHeap { .. }));
         prop_assert_eq!(
-            usize::from(program.header.heap_size_hint),
+            program.header.heap_size_hint as usize,
             store_count,
             "heap_size_hint must equal the number of distinct StoreHeap emissions"
         );
